requirements.txt
python-dotenv~=1.0.1
youtube-transcript-api~=0.6.3
deepmultilingualpunctuation~=1.0.1
openai~=1.59.7